# BUG/ENH: Ensure einsum(optimize=True) dispatches tensordot using ordered axes