@eric-wieser 2018-09-11
On my machine the difference is only a factor of two.
```
In [66]: %timeit np.tensordot(x, x, axes=((1, 2, 0), (1, 2, 0)))
21.3 ms ± 2.02 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [67]: %timeit np.tensordot(x, x, axes=((0, 1, 2), (0, 1, 2)))  # "better"
9.94 ms ± 1.23 ms per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

But note that you get a different result if the ordering of `xt` is different:

```

In [77]: xt = np.moveaxis(x, -1, 0)

In [78]: %timeit np.tensordot(xt, xt, axes=((0, 1, 2), (0, 1, 2)))  # uh oh - this is way worse now!
18.8 ms ± 103 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [79]: %timeit np.tensordot(xt, xt, axes=((1, 2, 0), (1, 2, 0)))
586 µs ± 39.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```