@eric-wieser 2018-09-11
> Should I close this and open an issue? What is your recommendation?

I think that would make sense. You're right that using a set here is causing a non-deterministic order on python < 3.6, which I'd still consider a bug.