@rsokl 2018-09-11
Okay, I will open an issue reflecting our discussion about `tensordot`.

> [...] is causing a non-deterministic order on python < 3.6 [...]

It is my understanding that sets still have random iteration order in Python 3.6 and beyond.:

![image](https://user-images.githubusercontent.com/29104956/45379355-d4540780-b5cd-11e8-8cb6-2ecbd5c9aa6d.png)
![image](https://user-images.githubusercontent.com/29104956/45379743-eb472980-b5ce-11e8-97ad-cb68c45bc2b9.png)

My understanding is that it is only dictionaries that are guaranteed to be ordered now.

I'll open a separate PR - as a bug fix - to ensure consistent axis-ordering for `einsum`.